split autosuggest into own component file
